fix: remove deleted uploads instantly
