+++ src/pages/MyUploads/hooks/useJobs.tsx
@@ -43,7 +43,6 @@ export default function useJobs(
   }
 
   useEffect(() => {
-
     fetchJobs();
   }, [backend]);
